landing page progress
