+++ FrontEnd/src/components/LandingPage/LandingPage.module.scss
@@ -1,0 +1,91 @@
+
+
+  .ellipseButton {
+    width: 54px;
+    height: 54px;
+    padding: 9.2px 21.1px 8.2px 20.1px;
+    background-color: #e7936d;
+  }
+  
+  .infoSectionWrapper {
+    margin-bottom: 1.5rem;
+  }
+  
+  .title{
+    margin: 10rem;
+  }
+
+  .intro{
+    margin: 10rem;
+    justify-content: center;
+  }
+  
+  .subtitle {
+    color: $primary-dark-color;
+    font-size: $lg-font-size;
+    margin: auto 0 auto 50px;
+  }
+  
+  .caption {
+    color: $primary-light-color;
+    font-size: 20px;
+    font-weight: bold;
+    justify-content: center;
+    margin: 1.25rem;
+  }
+  
+  .text {
+    color: rgb(46, 46, 46);
+    font-size: 14px;
+    margin-top: 2rem;
+    justify-content: center;
+  }
+  
+  .textLg {
+    color: rgb(46, 46, 46);
+    font-size: 20px;
+    justify-content: center;
+  }
+
+  .subtext {
+    color: $primary-dark-color;
+    font-size: 14px;
+    justify-content: center;
+    margin-bottom: 2rem;
+    margin-top: 0.25rem;
+  }
+
+  .subtextIntro {
+    color: rgb(46, 46, 46);
+    font-size: 14px;
+    justify-content: center;
+    margin-bottom: 2rem;
+    margin-top: 1rem;
+  }
+  
+  .textSm {
+    color: rgb(46, 46, 46);
+    font-size: 12px;
+  }
+  
+  .smallImg {
+    height: 100px;
+  }
+  
+  .buttonInner {
+    margin: inherit;
+  }
+  
+  .smallRow {
+    justify-content: center;
+    margin-bottom: 0.5rem;
+  }
+  
+  .bigRow, .center{
+    justify-content: center;
+  }
+
+  .iconRow{
+    margin-top: 3rem;
+    margin-bottom: 3rem;
+  }
