+++ FrontEnd/src/components/LandingPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from './LandingPage';
+export * from './LandingPage';
